requirements.txt
dspy>=2.6
playwright>=1.49
python-dotenv
